Crate tokio_net

Event loop that drives Tokio I/O resources.

The reactor is the engine that drives asynchronous I/O resources (like TCP and UDP sockets). It is backed by mio and acts as a bridge between mio and futures.

The crate provides:

Reactor is the main type of this crate. It performs the event loop logic.

Handle provides a reference to a reactor instance.

Registration and PollEvented allow third parties to implement I/O resources that are driven by the reactor.

Application authors will not use this crate directly. Instead, they will use the tokio crate. Library authors should only depend on tokio-net if they are building a custom I/O resource.

For more details, see reactor module documentation in the Tokio crate.

Modules

driver
Event loop that drives Tokio I/O resources.

tcp
TCP bindings for tokio.

udp
UDP bindings for tokio.

uds
Unix Domain Sockets for Tokio.

util
Utilities for implementing networking types.

Traits

ToSocketAddrs
Convert or resolve without blocking to one or more SocketAddr values.