Module tokio_net::udp::split

UdpSocket split support.

The split method splits a UdpSocket into a receive half and a send half, which can be used to receive and send datagrams concurrently, even from two different tasks.

The halves provide access to the underlying socket, implementing AsRef<UdpSocket>. This allows you to call UdpSocket methods that takes &self, e.g., to get local address, to get and set socket options, to join or leave multicast groups, etc.

The halves can be reunited to the original socket with their reunite methods.

Structs

ReuniteError
Error indicating two halves were not from the same socket, and thus could not be reunited.

UdpSocketRecvHalf
The recv half after split.

UdpSocketSendHalf
The send half after split.